tech-debt(backend): extract BaselinePricingSource base for 5 pricing sources (#12687)
anthropic/openai/deepseek/google/vertexai pricing sources each duplicated an
identical fetch() loop, differing only by the module-level _BASELINE table
and _PROVIDER constant. Added BaselinePricingSource in sources.py holding the
single fetch() implementation (handles both the 3-tuple and Anthropic's
4-tuple-with-cache_read baseline rows), and refactored all 5 sources to
subclass it, keeping only their _BASELINE/_PROVIDER data.
