+++ autobot-backend/llm_shared/pricing/openai_source.py
@@ -11,10 +11,7 @@ accept a future live-fetch implementation without changes to callers.
 
 from __future__ import annotations
 
-from autobot_shared.logging_manager import get_logger
-from llm_shared.pricing.sources import ModelPricing, PricingSource
-
-logger = get_logger(__name__)
+from llm_shared.pricing.sources import BaselinePricingSource
 
 _PROVIDER = "openai"
 
@@ -37,19 +34,7 @@ _BASELINE: list[tuple[str, float, float]] = [
 ]
 
 
-class OpenAIPricingSource(PricingSource):
+class OpenAIPricingSource(BaselinePricingSource):
     provider = _PROVIDER
-
-    async def fetch(self) -> dict[str, ModelPricing]:
-        now = self._now()
-        result: dict[str, ModelPricing] = {}
-        for model_id, inp, out in _BASELINE:
-            result[model_id] = ModelPricing(
-                provider=_PROVIDER,
-                model_id=model_id,
-                input_per_1m=inp,
-                output_per_1m=out,
-                updated_at=now,
-            )
-        logger.debug("OpenAIPricingSource.fetch returned %d models", len(result))
-        return result
+    _PROVIDER = _PROVIDER
+    _BASELINE = _BASELINE
